Backend/requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.0
bcrypt==3.2.0
cached-property==2.0.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
coloredlogs==15.0.1
cryptography==43.0.1
defusedxml==0.7.1
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
et_xmlfile==2.0.0
exchangelib==5.5.0
fastapi==0.115.2
h11==0.14.0
httptools==0.6.1
humanfriendly==10.0
idna==3.10
isodate==0.7.2
Jinja2==3.1.4
lxml==5.3.0
Mako==1.3.6
MarkupSafe==3.0.1
numpy==2.1.3
oauthlib==3.2.2
openpyxl==3.1.5
pandas==2.2.3
paramiko==3.5.0
passlib==1.7.4
psutil==6.1.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyNaCl==1.5.0
pyodbc==5.1.0
pyOpenSSL==24.2.1
#pysftp==0.2.9
pyspnego==0.11.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.12
python3-saml==1.16.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
requests_ntlm==1.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.45
starlette==0.40.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
uvicorn==0.31.1
watchfiles==0.24.0
websockets==13.1
xlrd==2.0.1
xmlsec==1.3.14
